requirements.txt
fastapi
uvicorn[standard]
joycon-python
hidapi
pyglm
cython-hidapi